Added in content
Added in the index.html. Added in the mc_forms.css. Added in the mc_styles.css. Linked the css sheets and the js sheet to the index. Added in the base content for the html page. Formatted the base html page in the mc_styles.css. Added the controls to the index page. Added in the selection list with states. All 50 states were put in that selection list. The submit buttons were added so that the form can be submitted, cancelled, or a special command button. Added in the css styles to style the form in the mc_forms.css file. Added in mobile styles so that the form works on mobile. Added in form validation.