Added in Styling
Changed the default style for the input boxes. Changed the default style for the textarea. Changed the defualt style for the selects. Added in more padding and margins around some of the elements. Fixed the navigation menu going offscreen on mobile. Added a function in the script to handle when the contact me button was pressed (Temporary). Removed the box around the input boxes. Fixed spacing in the footer. Buttons now scale properly on mobile. The text area now resizes with the screen only.